reorganization of uTests
